==> project shifting/code/3-shifting-plot.R <==
# ==========================================================================
# Plot: Choice Predictions from a Shifting Model
# Shifting between PH Heuristic (Korn & Bach, 2018) and Optimal Model (Houston & McNamara, 1988)
# ==========================================================================

# Load Packages ----------------------------------------------------------------
pacman::p_load(data.table, ggplot2, patchwork)
library(cognitivemodels)
windowsFonts(Arial=windowsFont("Arial"))

# Source -----------------------------------------------------------------------
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))
source("2-shifting-select.R")
#source("2-shifting-select1.R")

# Plot -------------------------------------------------------------------------
jitter <- position_jitter(width = 0.1)
stimuli[, model := as.factor(model)]
stimuli[,prhv := prhv * 100]

# # Median Shifting + PH & OPT Point Predictions--------------------------------
plotpredictions <- function(i){
  
  agg <- stimuli[nr == unique(nr)[i], .(median = median(prhv)), by = c("model","trial")]
  
  plot <- ggplot(stimuli[nr == unique(nr)[i] & model != "shift"] , 
                 aes(x = trial, y=prhv, shape = model))+
    theme_classic(base_family = "Arial") +
    geom_hline(yintercept = 0.5, linetype = 2, size = 0.5, alpha = 0.1)+
    geom_point(size = 3, position = jitter, alpha = 0.5, show.legend = FALSE) +
    scale_shape_manual(values=c(1,2,3),name = "Model",labels = c("Probability Heuristic", "Optimal Model", "Shifting Model (c=3)"))+ # Irgendwie verdreht (????)
    scale_colour_manual(values=c("black","green")) +
    #scale_fill_manual(values=c("grey","black"), name = "Model",labels = c("OPT softmax", "PH softmax"))+
    ylim(0,100) +
    scale_y_continuous(limits = c(0,100), expand = c(0,0)) +
    labs(x = 'Trial', y = '% Risky Choices')+
    #ggtitle(expression(paste(bold("Figure 2."), italic(" Model Predictions"))))+
    geom_point(agg[model == "shift"], mapping = aes(x = trial, y = median, shape = model), colour = "blue", inherit.aes = F, size = 4, stroke = 2)+
    geom_line(agg[model == "shift"], mapping = aes(x = trial, y = median, group = 1), linetype = 2, inherit.aes = F, colour = "blue", size = 0.8, alpha = 0.5)+
    theme(legend.background = element_rect(linetype = 1, size = 0.01, colour = "transparent", fill = "transparent"),
          legend.position = c(0.25 , 0.20),
          legend.title = element_text(size=12),legend.text = element_text(size = 11),
          title = element_text(size=13, family = "Arial"),
          text = element_text(family = "Arial", size = 11),
          axis.title.y = element_text(family = "Arial", size = 13),
          axis.title.x = element_text(family = "Arial", size = 13), 
          axis.text.y.left = element_text(family = "Arial", size = 11),
          axis.text.x.bottom = element_text(family = "Arial", size = 11),
          strip.background = element_rect(colour="black", size = 0.01),
          strip.text = element_text(family = "Arial")) + 
    guides(shape = guide_legend(override.aes = list(alpha = 1, colour = "black", stroke = 1)))
  print(plot)
  return(plot)
}


s1_opt_ph <- plotpredictions(1)
# ggsave("../figures/temp_shifting_a.png",width = 8, height = 4)
# s2_opt_ph <- plotpredictions(2)
# ggsave("../figures/temp_shifting_b.png",width = 8, height = 6)
# s3_opt_ph <- plotpredictions(3)
# ggsave("../figures/temp_shifting_c.png",width = 5, height = 4, scale = 0.9)


# comp <- s1_opt_ph  + plot_spacer() + s2_opt_ph  +
#   plot_layout(guides = "collect", widths = c(.4,.05,.4))
# ggsave("../figures/comp.png",width = 11, height = 4)

# Plot Median Predictions  -----------------------------------------------------------------------

plotmedian <- function(i){
  
  agg <- stimuli[nr == unique(nr)[i], .(median = median(prhv)), by = c("model","trial")]
  
  plot <- ggplot(agg[ model != "shift"] , 
                 aes(x = trial, y=median, shape = model, fill = model))+
    theme_classic(base_family = "Arial") +
    geom_hline(yintercept = 0.5, linetype = 2, size = 0.5, alpha = 0.1)+
    geom_point(size = 4, alpha = 0.5, fill = "grey", show.legend = FALSE) +
    scale_shape_manual(values=c(21,24,3),name = "Model",labels = c("Probability Heuristic", "Optimal Model", "Shifting Model (c=3)"))+ # Irgendwie verdreht (????)
    scale_colour_manual(values=c("black","green")) +
   # scale_fill_manual(values=c("grey","grey","blue"), name = "Model",labels = c("Probability Heuristic", "Optimal Model", "Shifting Model (c=3)"))+
    ylim(0,100) +
    scale_y_continuous(breaks = seq(0,100, 25), limits = c(0,110), expand = c(0,0)) +
    labs(x = 'Trial', y = '% Risky Choices',tag = 'b')+
    #ggtitle(expression(paste(bold("Figure 2."), italic(" Model Predictions"))))+
    geom_point(agg[model == "shift"], mapping = aes(x = trial, y = median, shape = model), colour = "blue", inherit.aes = F, size = 4, stroke = 2)+
    geom_line(agg[model == "shift"], mapping = aes(x = trial, y = median, group = 1), linetype = 2, inherit.aes = F, colour = "blue", size = 0.8, alpha = 0.5)+
    theme(legend.background = element_rect(linetype = 1, size = 0.01,  colour = "transparent", fill = "transparent"),
          legend.position = c(0.40 , 0.20),
          legend.title = element_text(size=13),legend.text = element_text(size = 12),
          title = element_text(size=13, family = "Arial"),
          text = element_text(family = "Arial", size = 11),
          axis.title.y = element_text(family = "Arial", size = 15),
          axis.title.x = element_text(family = "Arial", size = 15), 
          axis.text.y.left = element_text(family = "Arial", size = 12),
          axis.text.x.bottom = element_text(family = "Arial", size = 12),
          strip.background = element_rect(colour="black", size = 0.01),
          strip.text = element_text(family = "Arial")) + 
    guides(shape = guide_legend(override.aes = list(alpha = 1, colour = "black", stroke = 1)))
  print(plot)
  return(plot)
}


s3_opt_ph <- plotmedian(3)
ggsave("../figures/temp_shifting_c.png",width = 5, height = 4, scale = 0.9) # Used in proposal

plotmedian1 <- function(i){
  
  agg <- stimuli[nr == unique(nr)[i], .(median = median(prhv)), by = c("model","trial")]
  
  plot <- ggplot(agg[ model != "shift"] , 
                 aes(x = trial, y=median, shape = model, fill = model))+
    theme_classic(base_family = "Arial") +
    geom_hline(yintercept = 0.5, linetype = 2, size = 0.5, alpha = 0.1)+
    geom_point(size = 4, alpha = 0.5, fill = "grey", show.legend = FALSE) +
    scale_shape_manual(values=c(21,24,3),name = "",labels = c("Probability Heuristic", "Optimal Model", "Shifting Model"))+ # Irgendwie verdreht (????)
    scale_colour_manual(values=c("black","green")) +
    # scale_fill_manual(values=c("grey","grey","blue"), name = "Model",labels = c("Probability Heuristic", "Optimal Model", "Shifting Model (c=3)"))+
    ylim(0,100) +
    scale_y_continuous(breaks = c(0,50,100), limits = c(0,110), expand = c(0,0)) +
    labs(x = 'Choices (1-5)', y = '% Risky Choices')+
    #ggtitle(expression(paste(bold("Figure 2."), italic(" Model Predictions"))))+
    geom_point(agg[model == "shift"], mapping = aes(x = trial, y = median, shape = model), colour = "blue", inherit.aes = F, size = 4, stroke = 2)+
    geom_line(agg[model == "shift"], mapping = aes(x = trial, y = median, group = 1), linetype = 2, inherit.aes = F, colour = "blue", size = 0.8, alpha = 0.5)+
    theme(legend.background = element_rect(linetype = 1, size = 0.01,  colour = "transparent", fill = "transparent"),
          #legend.position = c(0.40 , 0.20),
          legend.position = "bottom",legend.direction = "horizontal",
          legend.title = element_text(size=13),
          legend.text = element_text(size = 12),
          title = element_text(size=13, family = "Arial"),
          text = element_text(family = "Arial", size = 11),
          axis.title.y = element_text(family = "Arial", size = 15),
          axis.title.x = element_text(family = "Arial", size = 15), 
          axis.text.y.left = element_text(family = "Arial", size = 12),
          axis.text.x.bottom =  element_blank(), #, element_text(family = "Arial", size = 12),
          axis.ticks.x.bottom=element_blank(),
          strip.background = element_rect(colour="black", size = 0.01),
          strip.text = element_text(family = "Arial"),
          plot.background = element_rect(fill = "#FBFBFB", colour = NA),
          panel.background = element_rect(fill = "#FBFBFB", colour = NA)) + 
    guides(shape = guide_legend(override.aes = list(alpha = 1, colour = "black", stroke = 1)))
  print(plot)
  return(plot)
}
stimuli[nr == unique(nr)[3]]
s3_opt_ph <- plotmedian1(3)
ggsave("../figures/temp_shifting_sed.png",width = 6, height = 5, scale = 0.91) # used for sed conf
# Shifting -------------------------------------------------------------------------

plotallpreds <- function(i){
  plot <- ggplot(stimuli[nr == unique(nr)[i]], 
                 aes(x = trial, y=prhv, fill = model))+
    theme_classic(base_size = 18, base_family = "Arial") +
    geom_hline(yintercept = 0.5, linetype = 2, size = 0.3)+
    geom_point(size = 1.4, position = jitter, shape = 21) +
    scale_fill_manual(values=c("blue","red", "green"), name = "Model",labels = c("OPT softmax", "PH softmax", "Shifting"))+
    scale_y_continuous(breaks = seq(0,100, 25), limits = c(0,110), expand = c(0,0)) +
    labs(x = 'Trial', y = '% Risky Choices', fill = "Model")+
    theme(legend.background = element_rect(linetype = 1, size = 0.01, colour = "black"),
          #legend.text = element_text(size = 13),
          text = element_text(family = "Arial", size = 11), axis.title.y = element_text(family = "Arial"),
          axis.title.x = element_text(family = "Arial"), 
          axis.text.y.left = element_text(family = "Arial"),
          axis.text.x.bottom = element_text(family = "Arial"),
          strip.background = element_rect(colour="black", size = 0.01),strip.text = element_text(family = "Arial")) +
    labs(title = "A", subtitle = paste0("Reach ", stimuli[nr == unique(stimuli$nr)[i], b[1]], " in 5 trial" ))
  # ggtitle( paste0("Stimuli ", unique(stimuli$nr)[i], ". Risky: ", stimuli[nr == unique(stimuli$nr)[i], xh[1]],
  #                 " (", stimuli[nr == unique(stimuli$nr)[i], pxh[1]], "); ", stimuli[nr == unique(stimuli$nr)[i], yh[1]],
  #                 " /  Safe: ", stimuli[nr == unique(stimuli$nr)[i], xl[1]],
  #                 " (", stimuli[nr == unique(stimuli$nr)[i], pxl[1]], "); ", stimuli[nr == unique(stimuli$nr)[i], yl[1]],". Budget: ",
  #                 stimuli[nr == unique(stimuli$nr)[i], b[1]], ". Difficulty: ", stimuli[nr == unique(stimuli$nr)[i], dhbin[1]]))
  return(plot)
}


s1 <- plotallpreds(1)
# ggsave("../figures/temp_shifting_a1.png",width = 8, height = 6)
# s2 <- plotallpreds(2)
# ggsave("../figures/temp_shifting_b1.png",width = 8, height = 6)
# s3 <- plotallpreds(3)
# ggsave("../figures/temp_shifting_c1.png",width = 8, height = 6)


# Nur Shifting ------------------------------------------------------------------------
plotshifting <- function(i){
  
  agg <- stimuli[nr == unique(nr)[i], .(median = median(prhv)), by = c("model","trial")]
  
  plot <- ggplot(stimuli[nr == unique(nr)[i] & model != "shift"] , 
                 aes(x = trial, y=prhv, shape = model))+
    theme_classic() +
    geom_point(size = 3, position = jitter, alpha = 0.5, show.legend = FALSE, colour="black")  +
    geom_point(agg[model == "shift"], mapping = aes(x = trial, y = median, shape = model), colour = "blue", inherit.aes = F, size = 4, stroke = 2)+
    geom_line(agg[model == "shift"], mapping = aes(x = trial, y = median, group = interaction(model), colour = model), 
              linetype = 2, inherit.aes = F, size = 0.8, alpha = 0.5, colour = "blue") +
    geom_point(agg[model != "shift"], mapping = aes(x = trial, y = median, shape = model), colour = "black", inherit.aes = F)+
    geom_line(agg[model != "shift"], mapping = aes(x = trial, y = median, group = interaction(model)), 
              linetype = 2, inherit.aes = F, size = 0.1, alpha = 0.8, colour = "grey") +
    scale_shape_manual(values=c(1,2,3,4,8),name = "Model",labels = c("Optimal Model", "Probability Heuristic", "Shifting Model")) +
    scale_colour_manual(values=c("black","black", "blue"),name = "Model",labels = c("Optimal Model", "Probability Heuristic", "Shifting Model")) +
    #scale_fill_manual(values=c("grey","black"), name = "Model",labels = c("OPT softmax", "PH softmax"))+
    scale_y_continuous(breaks = seq(0,100, 25), limits = c(0,110), expand = c(0,0)) +
    labs(x = 'Trial', y = 'Proportion of Risky Choices')+
    theme(legend.background = element_rect(linetype = 1, size = 0.01, colour = "black"),legend.text = element_text(size = 10),
          text = element_text(family = "Arial", size = 11), axis.title.y = element_text(family = "Arial", size = 13),
          axis.title.x = element_text(family = "Arial", size = 13), 
          axis.text.y.left = element_text(family = "Arial", size = 11),
          axis.text.x.bottom = element_text(family = "Arial", size = 11),
          strip.background = element_rect(colour="black", size = 0.01),strip.text = element_text(size = 11, family = "Arial"))+
    guides(shape = guide_legend(override.aes = list(alpha = 1, colour = "black", stroke = 1)),
           colour = guide_legend(override.aes = list(alpha = 1, colour = "black", stroke = 1))
           #fill = guide_legend(override.aes = list(alpha = 1)),
    ) +
    ggtitle(
      paste0("Stimuli ", unique(stimuli$nr)[i], ". Risky: ", stimuli[nr == unique(stimuli$nr)[i], xh[1]],
             " (", stimuli[nr == unique(stimuli$nr)[i], pxh[1]], "); ", stimuli[nr == unique(stimuli$nr)[i], yh[1]],
             " /  Safe: ", stimuli[nr == unique(stimuli$nr)[i], xl[1]],
             " (", stimuli[nr == unique(stimuli$nr)[i], pxl[1]], "); ", stimuli[nr == unique(stimuli$nr)[i], yl[1]],". Budget: ",
             stimuli[nr == unique(stimuli$nr)[i], b[1]], ". Difficulty: ", stimuli[nr == unique(stimuli$nr)[i], dhbin[1]]))
  print(plot)
  return(plot)
}

#
# 
 s1_opt_ph <- plotshifting(1)
# 
# ggsave("../figures/temp_shifting_a.png",width = 8, height = 6)
# s2_opt_ph <- plotfunc(2)
# ggsave("../figures/temp_shifting_b.png",width = 8, height = 6)
# s3_opt_ph <- plotfunc(3)
# ggsave("../figures/temp_shifting_c.png",width = 8, height = 6)
